Replace Combine’s Deferred { Future { } } with SQLiteFuture
Combine’s `Future` publisher appears to create memory leaks when used, at least on iOS 13. Additionally, it tries to resolve its promise as soon as it’s created, which is not what we want. So, we replaced it with SQLiteFuture to avoid memory issues and to get the behavior we want.