Concurrency (#42)
This pull request is a pretty dramatic change. It replaces the majority of the library's internal implementation with GRDB, while keeping the vast majority of the library's external API and behavior constant.

At this point, GRDB is simply much better maintained than this library. It doesn't make sense to reimplement the features GRDB has already implemented in this library. However, SQLite exposes a much simpler interface than GRDB—one which I personally find valuable to maintain.